Finished the constructors for Board class
The Board class now have 2 constructors, one without arguemnts which is
default 2-by-2 in size. The other one is to be defined, by setting row and col.

Class Board now have a class variable of type list of lists, where the innermost
lists contains objects of type Boxes.

Added a shell to a function, to print current state of the board in terminal window.
Not started with this function yet though.

Signed-off-by: Erik Henriksson <erikhenriksson91@gmail.com>